requirements.txt
Django~=5.0
djangorestframework

# Docs
drf-yasg
